+++ src/main/java/com/assessment/api/services/impl/PhotosServiceImpl.java
@@ -46,14 +46,20 @@ public class PhotosServiceImpl implements PhotosService {
 
 	@Override
 	public PhotosDTO getPhotosById(Integer id) {
+		if (null == id) {
+			return null;
+		}
 		Optional<Photos> Photos = photosRepository.findById(id);
 		return photosMapper.toDto(Photos.get());
 	}
 
 	@Override
 	public List<PhotosDTO> getPhotosByAlbumId(Integer albumId) {
+		if (null == albumId) {
+			return null;
+		}
 		Albums album = albumsRepository.findById(albumId).get();
-		Optional<List<Photos>> Photos = photosRepository.findByAlbumId(album);
+		Optional<List<Photos>> Photos = photosRepository.findByAlbum(album);
 		if (Photos.isPresent()) {
 			return mapEntityToDtoList(Photos.get());
 		}
@@ -62,6 +68,9 @@ public class PhotosServiceImpl implements PhotosService {
 
 	@Override
 	public List<PhotosDTO> getPhotoByTitle(String title) {
+		if (null == title) {
+			return null;
+		}
 		Optional<List<Photos>> Photos = photosRepository.findByTitle(title);
 		if (Photos.isPresent()) {
 			return mapEntityToDtoList(Photos.get());
@@ -71,6 +80,9 @@ public class PhotosServiceImpl implements PhotosService {
 
 	@Override
 	public List<PhotosDTO> getPhotoByURL(String url) {
+		if (null == url) {
+			return null;
+		}
 		Optional<List<Photos>> Photos = photosRepository.findByURL(url);
 		if (Photos.isPresent()) {
 			return mapEntityToDtoList(Photos.get());
@@ -80,6 +92,9 @@ public class PhotosServiceImpl implements PhotosService {
 
 	@Override
 	public List<PhotosDTO> getPhotoByThumbnailURL(String thumbnailUrl) {
+		if (null == thumbnailUrl) {
+			return null;
+		}
 		Optional<List<Photos>> Photos = photosRepository.findByThumbnailURL(thumbnailUrl);
 		if (Photos.isPresent()) {
 			return mapEntityToDtoList(Photos.get());
@@ -97,22 +112,27 @@ public class PhotosServiceImpl implements PhotosService {
 		photo.setTitle(title);
 		photo.setUrl(url);
 		photo.setThumbnailUrl(thumbnailUrl);
-		Albums album = albumsRepository.findById(albumId).get();
-		photo.setAlbumId(album);
+		if (null != albumId) {
+			Albums album = albumsRepository.findById(albumId).get();
+			photo.setAlbum(album);
+		}
 		photosRepository.save(photo);
 	}
 
 	@Override
 	public void updatePhoto(Integer id, Integer albumId, String title, String url, String thumbnailUrl) {
-		Photos photo = photosRepository.findById(id).get();
-		savePhoto(albumId, title, url, thumbnailUrl, photo);
+		if (null != id) {
+			Photos photo = photosRepository.findById(id).get();
+			savePhoto(albumId, title, url, thumbnailUrl, photo);
+		}
 	}
 
 	@Override
 	public void deletePhoto(Integer id) {
-		Photos Photo = photosRepository.findById(id).get();
-		photosRepository.delete(Photo);
-
+		if (null != id) {
+			Photos Photo = photosRepository.findById(id).get();
+			photosRepository.delete(Photo);
+		}
 	}
 
 }
